Add NCMBBase class
